+++ src/main/java/com/metamong/metaticket/domain/user/dto/UserDTO.java
@@ -43,10 +43,10 @@ public class UserDTO {
         @NotBlank(message = "이름은 필수 입력 값입니다.")
         private String name;
 
-        @NotEmpty(message = "나이는 필수 입력 값입니다.")
+        //@NotBlank(message = "나이는 필수 입력 값입니다.")
         private int age;
 
-        @NotEmpty(message = "전화번호는 필수 입력 값입니다.")
+        @NotBlank(message = "전화번호는 필수 입력 값입니다.")
         private String number;
     }
