+++ src/main/resources/templates/user/signup.html
@@ -13,7 +13,7 @@
 
     <title>SignUp</title>
 </head>
-<script src="/js/user/signup2.js"></script>
+<script src="/js/user/signup.js"></script>
 <script src="https://code.jquery.com/jquery-latest.min.js"></script>
 
 
@@ -91,41 +91,41 @@
 
                     <div class="userInfo">
                         <span class="instruction">이메일</span>
-                        <input class="userId" id="email" name="email" type="text" placeholder="6~20자 영문,숫자">
+                        <input class="userId" id="email" name="email" type="text" placeholder="이메일을 입력해주세요">
                         <input class="btn" id="emailck" type="button" value="중복확인">
                     </div>
                     <div id="emailmsg"></div>
 
                     <div class="userInfo">
                         <span class="instruction">비밀번호</span>
-                        <input class="userPw" id="passwd" name="passwd" type="password" placeholder="8~12자 영문,숫자,특수문자">
+                        <input class="userPw" id="passwd" name="passwd" type="password" placeholder="8~12자의 영문,숫자,특수문자로 이루어진 비밀번호를 입력해주세요">
                     </div>
                     <div id="passwdmsg"></div>
 
                     <div class="userInfo">
                         <span class="instruction">비밀번호 확인</span>
-                        <input class="userPwChk" id="passwdck" type="password" placeholder="8~12자 영문,숫자,특수문자">
+                        <input class="userPwChk" id="passwdck" type="password" placeholder="비밀번호를 확인해주세요">
                     </div>
                     <div id="passwdckmsg"></div>
 
                     <div class="userInfo">
                         <span class="instruction">이름</span>
-                        <input class="userName" id="name" name="name" type="text">
+                        <input class="userName" id="name" name="name" type="text" placeholder="이름을 입력해주세요">
                     </div>
                     <div id="namemsg"></div>
 
                     <div class="userInfo">
                         <span class="instruction">생년월일</span>
-                        <input class="userBirth" id="age" name="age" type="text">
+                        <input class="userBirth" id="age" name="age" type="text" placeholder="6자로 이루어진 생년월일을 입력해주세요">
                     </div>
                     <div id="agemsg"></div>
 
                     <div class="userInfo">
-                        <span class="instruction">휴대폰</span>
-                        <input class="userPhone" id="number" name="number" type="text">
-                        <input th:if="${session.random==null}" class="btn" id="numberck" type="button" value="인증번호">
-                        <input th:unless="${session.random==null}" class="userPhone" id="authnumber" type="text">
-                        <input th:unless="${session.random==null}" class="btn" id="smsck" type="button" value="인증">
+                        <span class="instruction">휴대전화</span>
+                        <input class="userPhone" id="number" name="number" type="text" placeholder="휴대전화를 인증해주세요">
+                        <input class="btn" id="numberck" type="button" value="인증번호">
+                        <input class="userPhone" id="authnumber" type="hidden" placeholder="인증번호를 입력해주세요.">
+                        <input class="btn" id="smsck" type="hidden" value="인증">
                     </div>
                     <div id="numbermsg"></div>
